Extracted ArgumentException tests to files.
