+++ Tests/Serilog.Exceptions.Test/Destructurers/ArgumentNullExceptionTest.cs
@@ -1,0 +1,16 @@
+﻿namespace Serilog.Exceptions.Test.Destructurers
+{
+    using System;
+    using Xunit;
+    using static LogJsonOutputUtils;
+
+    public class ArgumentNullExceptionTest
+    {
+        [Fact]
+        public void ArgumentNullException_ParamNameIsAttachedAsProperty()
+        {
+            var argumentException = new ArgumentNullException("testParamName", "MSG");
+            Test_LoggedExceptionContainsProperty(argumentException, "ParamName", "testParamName");
+        }
+    }
+}
