+++ Tests/Serilog.Exceptions.Test/ExceptionDestructurerTest.cs
@@ -10,7 +10,7 @@
     using Serilog.Formatting;
     using Serilog.Formatting.Json;
     using Xunit;
-    using static Serilog.Exceptions.Test.Destructurers.LogJsonOutputUtils;
+    using static Destructurers.LogJsonOutputUtils;
 
     public class ExceptionDestructurerTest
     {
@@ -108,34 +108,6 @@
         {
             var applicationException = new ApplicationException();
             Test_LoggedExceptionContainsProperty(applicationException, "Type", "System.ApplicationException");
-        }
-
-        [Fact]
-        public void ArgumentException_ParamNameIsAttachedAsProperty()
-        {
-            var argumentException = new ArgumentException("MSG", "testParamName");
-            Test_LoggedExceptionContainsProperty(argumentException, "ParamName", "testParamName");
-        }
-
-        [Fact]
-        public void ArgumentNullException_ParamNameIsAttachedAsProperty()
-        {
-            var argumentException = new ArgumentNullException("testParamName", "MSG");
-            Test_LoggedExceptionContainsProperty(argumentException, "ParamName", "testParamName");
-        }
-
-        [Fact]
-        public void ArgumentOfOutRangeException_ParamNameIsAttachedAsProperty()
-        {
-            var argumentException = new ArgumentOutOfRangeException("testParamName");
-            Test_LoggedExceptionContainsProperty(argumentException, "ParamName", "testParamName");
-        }
-
-        [Fact]
-        public void ArgumentOfOutRangeException_ActualValueIsAttachedAsProperty()
-        {
-            var argumentException = new ArgumentOutOfRangeException("testParamName", "ACTUAL_VALUE", "MSG");
-            Test_LoggedExceptionContainsProperty(argumentException, "ActualValue", "ACTUAL_VALUE");
         }
 
         [Fact]
